Set up the graph classes from completed lab 3 structure
Copied over the files from the completed lab 3. Also brought in
the files from the assignment zip, which were the test cases.
Added a read from file method to the GraphTester class. This reads
from a file with the specified format and puts it into a new graph.
Started setting up dijkstras.